Update default bed Kp/Ki/Kd for Leenovo AC bed
Kd didn't change much, the bed is still the bed and has a lot of mass
to it, giving a fair bit of 'inertia'.  Ki drops to almost nothing as
the power of the bed means it can correct temperature quickly should
it drop below the set point.  Kp eases back an order of magnitude.